LAA-18 changed services, application.properties, added ExceptionHandler,  update test for StatusController, ResourceTypeController

+++ src/main/java/ru/aston/mcs/service/impl/ResourceTypeServicesImpl.java
@@ -2,6 +2,11 @@ package ru.aston.mcs.service.impl;
 
 import org.springframework.stereotype.Service;
 import ru.aston.mcs.dto.ResourceTypeDTO;
+import ru.aston.mcs.dto.StatusDTO;
+import ru.aston.mcs.entity.ResourceType;
+import ru.aston.mcs.entity.Status;
+import ru.aston.mcs.exception.EntityNotFoundException;
+import ru.aston.mcs.exception.InvalidRequestException;
 import ru.aston.mcs.mapper.ResourceTypeMapper;
 import ru.aston.mcs.repository.ResourceTypeRepository;
 import ru.aston.mcs.service.ResourceTypeService;
@@ -31,17 +36,39 @@ public class ResourceTypeServicesImpl implements ResourceTypeService {
 
         return typeResourcesMapper.toDTO(
                 typeResourcesRepository.findById(nameId)
-                        .orElseThrow(RuntimeException::new));
+                        .orElseThrow(() -> new EntityNotFoundException(nameId)));
     }
 
     @Override
-    public void addAndSaveResourceType(ResourceTypeDTO resourceTypeDTO) {
+    public void saveResourceType(ResourceTypeDTO resourceTypeDTO) {
 
         typeResourcesRepository.save(typeResourcesMapper.toModel(resourceTypeDTO));
     }
 
     @Override
+    public ResourceTypeDTO updateResourceType(ResourceTypeDTO resourceTypeDTO) {
+
+        if (resourceTypeDTO == null || resourceTypeDTO.getNameId() == null) {
+            throw new InvalidRequestException();
+        }
+
+        Long resourceTypeId = resourceTypeDTO.getNameId();
+        ResourceType resourceTypeFromDb =  typeResourcesRepository.findById(resourceTypeId)
+                .orElseThrow( () -> new EntityNotFoundException(resourceTypeId));
+
+        resourceTypeFromDb.setName(resourceTypeDTO.getName());
+        resourceTypeFromDb.setCost(resourceTypeDTO.getCost());
+
+        return typeResourcesMapper.toDTO(typeResourcesRepository.save(resourceTypeFromDb));
+    }
+
+    @Override
     public void deleteResourceType(Long nameId) {
+
+        if (typeResourcesRepository.findById(nameId).isEmpty()) {
+            throw new EntityNotFoundException( nameId );
+        }
+
         typeResourcesRepository.deleteById(nameId);
     }
